requirements.txt
streamlit==1.53.0
python-dotenv==1.2.1
google-generativeai==0.8.6